feat: Add tests to project
